removed redundant switch file in test
