+++ lib/render/dom/create/append.js
@@ -16,26 +16,29 @@ exports.static = function staticAppendNode (target, pnode, domNode) {
   pnode.appendChild(domNode)
 }
 
-exports.state = function stateAppendNode (target, pnode, domNode, subs, tree, uid, order) {
+exports.state = function stateAppendNode (target, pnode, node, subs, tree, uid, order) {
   var fragment
+
+  console.info('wtf is this', pnode, node)
+
   if (isFragment(pnode)) {
     fragment = pnode
     pnode = findParent(pnode)
   }
   if (order !== void 0) {
     const next = findNode(order, pnode)
-    domNode._order = order
+    node._order = order
     if (next) {
-      if (fragment) { fragment.push(domNode) }
-      pnode.insertBefore(domNode, next)
+      if (fragment) { fragment.push(node) }
+      pnode.insertBefore(node, next)
     } else {
       pnode._last = order
-      if (fragment) { fragment.push(domNode) }
-      pnode.appendChild(domNode)
+      if (fragment) { fragment.push(node) }
+      pnode.appendChild(node)
     }
   } else {
-    if (fragment) { fragment.push(domNode) }
-    pnode.appendChild(domNode)
+    if (fragment) { fragment.push(node) }
+    pnode.appendChild(node)
   }
 }
