+++ test/index.js
@@ -10,8 +10,9 @@
 // require('./parent')
 // require('./group')
 // require('./clone')
-require('./any')
+// require('./any')
 // require('./subscribe')
 // require('./fragment')
 // require('./reference')
 // require('./test')
+require('./switch')
