+++ test/switch/index.js
@@ -1,4 +1,5 @@
 'use strict'
-require('./branch')
-require('./nested')
-require('./property')
+// require('./branch')
+// require('./nested')
+// require('./property')
+require('./context')
